feat: Update API URLs to use production endpoints
The code changes update the API URLs in the getGenders.ts and getUsers.ts files to use the production endpoints (https://api.bledbonds.es/api/v1) instead of the local development server (http://localhost:3001/api/v1). This ensures that the application is making requests to the correct API URLs in a production environment.